requirements.txt
django==5.0.6
django-markdown==0.8.4
django-taggit==5.0.1